# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PostAWay API",
    "description": "API for social networking platform",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "JWT": { "in": "header", "name": "Authorization", "type": "apiKey" }
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get posts of logged-in user",
        "description": "Retrieves all posts created by the authenticated user",
        "security": [{ "JWT": {} }],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found!" }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create new post",
        "description": "Create a new post with caption and image",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": { "type": "string" },
                  "imageUrl": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "security": [{ "JWT": {} }],
        "responses": {
          "201": { "description": "OK" }
        }
      }
    },
...
  }
}